03-data-warehouse/homework/pyproject.toml
[project]
name = "homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastparquet>=2025.12.0",
    "google-cloud-storage>=3.8.0",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "requests>=2.32.5",
]
